Enhanced code through refactoring in aprgaudiomanipulator module

+++ AprgAudioManipulator/AprgAudioManipulator/src/AudioManipulator/NearestSamples.cpp
@@ -12,27 +12,25 @@ namespace AprgAudio {
 
 NearestSamples::NearestSamples(Samples const& samples) : m_samples(samples) { saveToValuesToIndexes(); }
 
-Indexes NearestSamples::getNearestSamplesIndexes(double const value, unsigned int const recommendedNumberOfSamples) {
+Indexes NearestSamples::getNearestSamplesIndexes(double const value, int const recommendedNumberOfSamples) {
     Indexes result;
     pair<ValuesToIndexes::const_iterator, ValuesToIndexes::const_iterator> lowerAndUpperIterator(
         containerHelper::getLowerAndUpperConstIteratorsInMap(m_valuesToIndexes, value));
 
-    set<unsigned int> indexesInOrder;
+    set<int> indexesInOrder;
     for (ValuesToIndexes::const_iterator it = lowerAndUpperIterator.first; it != lowerAndUpperIterator.second; it++) {
         indexesInOrder.emplace(it->second);
     }
     ValuesToIndexes::const_iterator itLower = lowerAndUpperIterator.first;
     ValuesToIndexes::const_iterator itUpper = lowerAndUpperIterator.second;
-    while (indexesInOrder.size() < recommendedNumberOfSamples && itLower != m_valuesToIndexes.cbegin() &&
-           itUpper != m_valuesToIndexes.cend()) {
+    while (static_cast<int>(indexesInOrder.size()) < recommendedNumberOfSamples &&
+           itLower != m_valuesToIndexes.cbegin() && itUpper != m_valuesToIndexes.cend()) {
         if (itLower != m_valuesToIndexes.cbegin()) {
             indexesInOrder.emplace(itLower->second);
-            itLower--;
-        }
+            itLower--;        }
         if (itUpper != m_valuesToIndexes.cend()) {
             indexesInOrder.emplace(itUpper->second);
-            itUpper--;
-        }
+            itUpper--;        }
     }
     result.reserve(result.size());
     copy(indexesInOrder.cbegin(), indexesInOrder.cend(), back_inserter(result));
@@ -40,12 +38,10 @@ Indexes NearestSamples::getNearestSamplesIndexes(double const value, unsigned in
 }
 
 void NearestSamples::saveToValuesToIndexes() {
-    unsigned int i = 0;
+    int i = 0;
     for (double const sample : m_samples) {
         m_valuesToIndexes.emplace(sample, i++);
-    }
-}
+    }}
 
 }  // namespace AprgAudio
-
 }  // namespace alba
